perf: Add functional submit form
The submit form currently has some form of validation on the front end,
e.g. it cannot book times that are already booked, it can only select
consecutive hours from the schedule, you have to provide a name, email,
and phone number.

The form creates the event in Google, along with a well formatted
description. It then redirects to a page confirming the booking.

Still need to work on some validation:
* You can currently submit without providing a booking time
* No phone number validation
* No SQL/injection attack protection
* No onscreen feedback if form details are incorrect
* Redirect screen ought to tell you what date was booked specifically

And of course, need to add payment options 😎

Could really probably do with some unit tests on all of this.

This guide will probably be useful:
https://www.thisdot.co/blog/handling-forms-in-svelte